requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
gunicorn==23.0.0
pydantic-settings==2.4.0
openai
pandas
asyncio
requests
pyjwt
aiohttp
locust
statistics
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2
python-dotenv==1.0.1
